+++ registry/users.go
@@ -2,15 +2,10 @@ package registry
 
 import (
 	"context"
-	"errors"
-	"fmt"
-	"log"
 	"net/url"
 
 	"github.com/manifoldco/torus-cli/apitypes"
 	"github.com/manifoldco/torus-cli/envelope"
-
-	"github.com/manifoldco/torus-cli/daemon/crypto"
 )
 
 // UsersClient represents the  registry `/users` endpoints.
@@ -20,7 +15,7 @@ type UsersClient struct {
 
 // Create attempts to register a new user
 func (u *UsersClient) Create(ctx context.Context, userObj *envelope.User,
-	signup apitypes.Signup) (*envelope.User, error) {
+	signup apitypes.Signup) (envelope.UserInf, error) {
 
 	v := &url.Values{}
 	if signup.InviteCode != "" {
@@ -33,19 +28,13 @@ func (u *UsersClient) Create(ctx context.Context, userObj *envelope.User,
 		v.Set("email", signup.Email)
 	}
 
-	user := envelope.User{}
-	err := u.client.RoundTrip(ctx, "POST", "/users", v, &userObj, &user)
+	e := envelope.Unsigned{}
+	err := u.client.RoundTrip(ctx, "POST", "/users", v, &userObj, &e)
 	if err != nil {
 		return nil, err
 	}
 
-	err = validateSelf(&user)
-	if err != nil {
-		log.Printf("Invalid user object: %s", err)
-		return nil, err
-	}
-
-	return &user, nil
+	return envelope.ConvertUser(&e)
 }
 
 // VerifyEmail will confirm the user's email with the registry
@@ -57,47 +46,12 @@ func (u *UsersClient) VerifyEmail(ctx context.Context, verifyCode string) error 
 }
 
 // Update patches the user object with whitelisted fields
-func (u *UsersClient) Update(ctx context.Context, userObj interface{}) (*envelope.User, error) {
-	user := envelope.User{}
-	err := u.client.RoundTrip(ctx, "PATCH", "/users/self", nil, userObj, &user)
+func (u *UsersClient) Update(ctx context.Context, userObj interface{}) (envelope.UserInf, error) {
+	e := envelope.Unsigned{}
+	err := u.client.RoundTrip(ctx, "PATCH", "/users/self", nil, userObj, &e)
 	if err != nil {
 		return nil, err
 	}
 
-	err = validateSelf(&user)
-	if err != nil {
-		log.Printf("Invalid user object: %s", err)
-		return nil, err
-	}
-
-	return &user, nil
-}
-
-func validateSelf(s *envelope.User) error {
-	if s.Version != 1 {
-		return errors.New("version must be 1")
-	}
-
-	if s.Body == nil {
-		return errors.New("missing body")
-	}
-
-	if s.Body.Master == nil {
-		return errors.New("missing master key section")
-	}
-
-	if s.Body.Master.Alg != crypto.Triplesec {
-		return &apitypes.Error{
-			Type: apitypes.InternalServerError,
-			Err: []string{
-				fmt.Sprintf("Unknown alg: %s", s.Body.Master.Alg),
-			},
-		}
-	}
-
-	if len(*s.Body.Master.Value) == 0 {
-		return errors.New("Zero length master key found")
-	}
-
-	return nil
+	return envelope.ConvertUser(&e)
 }